supervisor operations
